moving secret logic to secret package
